# Slightly reduce TTFX for `SessionActions.open`