@rikhuijzer 2022-05-09
In 82a2c13, the explicit `precompile` call for the method with signature is removed again
```julia
function load_notebook_nobackup(@nospecialize(io::IO), @nospecialize(path::AbstractString))::Notebook
```
This will be automated by https://github.com/rikhuijzer/PrecompileSignatures.jl/pull/9.